# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Compensation Strategy Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business-goals:
    get:
      operationId: listBusinessGoals
      tags: [businessGoals]
      summary: List all business goal profiles
      responses:
        "200":
          description: List of business goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessGoal"
    post:
      operationId: createBusinessGoal
      tags: [businessGoals]
      summary: Create a new business goal profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessGoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessGoal"
  /business-goals/{id}:
    get:
      operationId: getBusinessGoal
      tags: [businessGoals]
      summary: Get a business goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessGoal"
        "404":
          description: Not found
    patch:
      operationId: updateBusinessGoal
      tags: [businessGoals]
      summary: Update a business goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessGoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessGoal"
        "404":
          description: Not found
    delete:
      operationId: deleteBusinessGoal
      tags: [businessGoals]
      summary: Delete a business goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /business-goals/{id}/duplicate:
    post:
      operationId: duplicateBusinessGoal
      tags: [businessGoals]
      summary: Duplicate a business goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessGoal"

  # ─────────────────────────────────────────
  # Current Reality
  # ─────────────────────────────────────────
...
  /clinicians:
    get:
      operationId: listClinicians
      tags: [clinicians]
      summary: List all clinician profiles
      responses:
        "200":
          description: List of clinicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clinician"
    post:
      operationId: createClinician
      tags: [clinicians]
      summary: Create a clinician profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinician"
  /clinicians/{id}:
    get:
      operationId: getClinician
      tags: [clinicians]
      summary: Get a clinician by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clinician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinician"
        "404":
          description: Not found
    patch:
      operationId: updateClinician
      tags: [clinicians]
      summary: Update a clinician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicianUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinician"
        "404":
          description: Not found
    delete:
      operationId: deleteClinician
      tags: [clinicians]
      summary: Delete a clinician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /clinicians/{id}/duplicate:
    post:
      operationId: duplicateClinician
      tags: [clinicians]
      summary: Duplicate a clinician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated clinician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinician"

  # ─────────────────────────────────────────
  # Scenarios
  # ─────────────────────────────────────────
...
components:
  schemas:
...
    BusinessGoal:
      type: object
      required:
        - id
        - name
        - timeHorizon
        - ownerPayGoal
        - secondOwnerPayGoal
        - annualOverheadGoal
        - businessProfitGoal
        - buildingFundGoal
        - emergencyReserveGoal
        - growthFundGoal
        - desiredCliniciansCount
        - desiredOwnerClinicalCaseload
        - notes
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        timeHorizon:
          type: string
          enum: [1year, 3year, 5year, 10year]
        ownerPayGoal:
          type: number
        secondOwnerPayGoal:
          type: number
        annualOverheadGoal:
          type: number
        businessProfitGoal:
          type: number
        buildingFundGoal:
          type: number
        emergencyReserveGoal:
          type: number
        growthFundGoal:
          type: number
        desiredCliniciansCount:
          type: integer
        desiredOwnerClinicalCaseload:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    BusinessGoalInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        timeHorizon:
          type: string
          enum: [1year, 3year, 5year, 10year]
        ownerPayGoal:
          type: number
        secondOwnerPayGoal:
          type: number
        annualOverheadGoal:
          type: number
        businessProfitGoal:
          type: number
        buildingFundGoal:
          type: number
        emergencyReserveGoal:
          type: number
        growthFundGoal:
          type: number
        desiredCliniciansCount:
          type: integer
        desiredOwnerClinicalCaseload:
          type: integer
        notes:
          type: string
    BusinessGoalUpdate:
      type: object
      properties:
        name:
          type: string
        timeHorizon:
          type: string
          enum: [1year, 3year, 5year, 10year]
        ownerPayGoal:
          type: number
        secondOwnerPayGoal:
          type: number
        annualOverheadGoal:
          type: number
        businessProfitGoal:
          type: number
        buildingFundGoal:
          type: number
        emergencyReserveGoal:
          type: number
        growthFundGoal:
          type: number
        desiredCliniciansCount:
          type: integer
        desiredOwnerClinicalCaseload:
          type: integer
        notes:
          type: string

    # ─── Current Reality ─────────────────────
...
    Clinician:
      type: object
      required:
        - id
        - label
        - roleType
        - classification
        - sessionRate
        - sessionsPerWeek
        - weeksWorkedPerYear
        - preCapClinicianSplit
        - preCapPracticeSplit
        - capEnabled
        - capAmount
        - postCapClinicianSplit
        - postCapPracticeSplit
        - w2EmployerFicaPct
        - futaSutaPct
        - workersCompPct
        - otherEmployerBurdenPct
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        label:
          type: string
        roleType:
          type: string
          enum: [senior_clinician, associate, owner, contractor, other]
        classification:
          type: string
          enum: [w2, 1099, owner, other]
        sessionRate:
          type: number
        sessionsPerWeek:
          type: number
        weeksWorkedPerYear:
          type: number
        preCapClinicianSplit:
          type: number
        preCapPracticeSplit:
          type: number
        capEnabled:
          type: boolean
        capAmount:
          type: number
        postCapClinicianSplit:
          type: number
        postCapPracticeSplit:
          type: number
        w2EmployerFicaPct:
          type: number
        futaSutaPct:
          type: number
        workersCompPct:
          type: number
        otherEmployerBurdenPct:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ClinicianInput:
      type: object
      required:
        - label
      properties:
        label:
          type: string
        roleType:
          type: string
          enum: [senior_clinician, associate, owner, contractor, other]
        classification:
          type: string
          enum: [w2, 1099, owner, other]
        sessionRate:
          type: number
        sessionsPerWeek:
          type: number
        weeksWorkedPerYear:
          type: number
        preCapClinicianSplit:
          type: number
        preCapPracticeSplit:
          type: number
        capEnabled:
          type: boolean
        capAmount:
          type: number
        postCapClinicianSplit:
          type: number
        postCapPracticeSplit:
          type: number
        w2EmployerFicaPct:
          type: number
        futaSutaPct:
          type: number
        workersCompPct:
          type: number
        otherEmployerBurdenPct:
          type: number
        notes:
          type: string
    ClinicianUpdate:
      type: object
      properties:
        label:
          type: string
        roleType:
          type: string
          enum: [senior_clinician, associate, owner, contractor, other]
        classification:
          type: string
          enum: [w2, 1099, owner, other]
        sessionRate:
          type: number
        sessionsPerWeek:
          type: number
        weeksWorkedPerYear:
          type: number
        preCapClinicianSplit:
          type: number
        preCapPracticeSplit:
          type: number
        capEnabled:
          type: boolean
        capAmount:
          type: number
        postCapClinicianSplit:
          type: number
        postCapPracticeSplit:
          type: number
        w2EmployerFicaPct:
          type: number
        futaSutaPct:
          type: number
        workersCompPct:
          type: number
        otherEmployerBurdenPct:
          type: number
        notes:
          type: string

    # ─── Scenario ────────────────────────────
...