# lib/api-zod/src/generated/types/businessGoalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Compensation Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessGoalInputTimeHorizon } from './businessGoalInputTimeHorizon';

export interface BusinessGoalInput {
  name: string;
  timeHorizon?: BusinessGoalInputTimeHorizon;
  ownerPayGoal?: number;
  secondOwnerPayGoal?: number;
  annualOverheadGoal?: number;
  businessProfitGoal?: number;
  buildingFundGoal?: number;
  emergencyReserveGoal?: number;
  growthFundGoal?: number;
  desiredCliniciansCount?: number;
  desiredOwnerClinicalCaseload?: number;
  notes?: string;
}