# lib/api-zod/src/generated/types/clinician.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Compensation Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ClinicianClassification } from './clinicianClassification';
import type { ClinicianRoleType } from './clinicianRoleType';

export interface Clinician {
  id: number;
  label: string;
  roleType: ClinicianRoleType;
  classification: ClinicianClassification;
  sessionRate: number;
  sessionsPerWeek: number;
  weeksWorkedPerYear: number;
  preCapClinicianSplit: number;
  preCapPracticeSplit: number;
  capEnabled: boolean;
  capAmount: number;
  postCapClinicianSplit: number;
  postCapPracticeSplit: number;
  w2EmployerFicaPct: number;
  futaSutaPct: number;
  workersCompPct: number;
  otherEmployerBurdenPct: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}