# lib/api-zod/src/generated/types/clinicianInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Compensation Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ClinicianInputClassification } from './clinicianInputClassification';
import type { ClinicianInputRoleType } from './clinicianInputRoleType';

export interface ClinicianInput {
  label: string;
  roleType?: ClinicianInputRoleType;
  classification?: ClinicianInputClassification;
  sessionRate?: number;
  sessionsPerWeek?: number;
  weeksWorkedPerYear?: number;
  preCapClinicianSplit?: number;
  preCapPracticeSplit?: number;
  capEnabled?: boolean;
  capAmount?: number;
  postCapClinicianSplit?: number;
  postCapPracticeSplit?: number;
  w2EmployerFicaPct?: number;
  futaSutaPct?: number;
  workersCompPct?: number;
  otherEmployerBurdenPct?: number;
  notes?: string;
}