# lib/api-zod/src/generated/types/clinicianUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Compensation Strategy Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ClinicianUpdateClassification } from './clinicianUpdateClassification';
import type { ClinicianUpdateRoleType } from './clinicianUpdateRoleType';

export interface ClinicianUpdate {
  label?: string;
  roleType?: ClinicianUpdateRoleType;
  classification?: ClinicianUpdateClassification;
  sessionRate?: number;
  sessionsPerWeek?: number;
  weeksWorkedPerYear?: number;
  preCapClinicianSplit?: number;
  preCapPracticeSplit?: number;
  capEnabled?: boolean;
  capAmount?: number;
  postCapClinicianSplit?: number;
  postCapPracticeSplit?: number;
  w2EmployerFicaPct?: number;
  futaSutaPct?: number;
  workersCompPct?: number;
  otherEmployerBurdenPct?: number;
  notes?: string;
}